% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map Locations by State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state code that is in numeric or string format}

\item{year}{A single year in numeric or string format}
}
\value{
If the given state code exists and the filename of the given year exists, this function will return a map displaying
the state boundary and all accidents in that given year as points. If the state number does not exist within the dataset,
a message will print saying that the state number is invalid. If there are no observations for a given year with a state,
the function will return a message reporting that there are no accidents to plot.
}
\description{
This function will map the longitude and latitude of each observation within a given state and year.
}
\examples{
\dontrun{fars_map_state(53, 2013)}
\dontrun{fars_map_state(53, 2014)}
}
