% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that the function will search for.}
}
\value{
If the filename exists, this function returns the csv as a tibble/data.frame. If the file does not exist,
the function will instead will return with a message saying that the file does not exist.
}
\description{
This is a function that will search for and read-in a csv file based on a given filename. If the file exists, it will
read the csv file and convert it into a tibble/data.frame
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{fars_read("accident_2015.csv.bz2")}
}
