% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple fars files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing one or more years. The years can be numeric or string.}
}
\value{
If the file of a given year in the vector of years exists, then this function will return a list of
tibble/data.frames of columns MONTH and year from each file. If a file does not exist with one of the listed years,
then it will print a warning saying that that year is invalid and return a NULL object.
}
\description{
This function will read in multiple years worth of files and return a list of tibble/data.frames of just the columns
MONTH and year.
}
\examples{
\dontrun{fars_read_years(c(2014, 2015))}
\dontrun{fars_read_years(c(2014))}
}
