% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise the number of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing one or more years. The years can be numeric or string.}
}
\value{
This function returns a tibble/data.frame containing the number of observations for all given years for each month.
The given years are the column names while each row represents monthly counts.
}
\description{
This function will create a summary crosstabulating of the number of observations by month and year in a tibble/data.frame.
}
\examples{
\dontrun{fars_summarize_years(c(2014, 2015))}
\dontrun{fars_summarize_years(c(2014))}
}
