% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create the filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Consecutive digits in either string or numeric format that will be inserted to a filename string template}
}
\value{
This function returns a string
}
\description{
This function will take digits representing the year and insert it as part of a new string
to compose a filename.
}
\examples{
\dontrun{make_filename("2017")}
\dontrun{make_filename(2017)}
\dontrun{make_filename(17)}
}
