% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map showing accident location in a state on a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state to be plotted, as an integer. You can lookup the state codes
in the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812559}{FARS code book}.}

\item{year}{The year of interest}
}
\value{
None
}
\description{
This function plots the location of accidents in a given state for a given year,
using data read from the FARS database.
}
\note{
This function will fail on invalid states or years, and will create an empty plot
if no accidents were found.
}
\examples{
\dontrun{
fars_map_state(1, 2015)
}

}
