% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a FARS accident table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename to read}
}
\value{
A tibble containing the FARS accidents contained in the given file.
}
\description{
This function reads in a CSV file containing FARS accidents for a single year,
and returns a dplyr tibble. The function `stop`s if the file does not exist.
}
\examples{
fars_read(make_filename(2013))
}
