% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in the FARS accident reports for a range of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector of years}
}
\value{
A list of cleaned up dataframes containing the `MONTH` and the `year` of each accident
}
\description{
This function reads in the report accidents for the given years,
reads the corresponding files into tibbles, and keeps only the month and year of a given accident.
It then returns a list of the cleaned up data as a list.
}
\note{
An invalid year will generate a NULL element, and a warning message will be printed.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
}

}
