% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return a monthly summary of the FARS for a range of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector of years to compare}
}
\value{
A dataframe of the accident counts for each month of each request year.
There is a column per year, and a row for each month.
}
\description{
This function returns a monthly comparison of the number of accidents in each year.
}
\details{
The accident report CSV for each year is read using \link{fars_read_yars},
a summary of the counts is created using \link{dplyr::summarize},
and the data is then converted into a wide table for easier comparison using
\link{tidyr::spread}.
}
\note{
Invalid years will be ignored, and a warning message will be printed.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
