% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Returns the accident table's filename for a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The requested year as an integer}
}
\value{
A string containing the filename of the CSV containing the requested year's information
}
\description{
Returns the accident table's filename for a specific year
}
\examples{
\dontrun{
make_filename(2013)
}
}
